+++ database/migrations/2024_02_08_000000_version_701.php
@@ -14,6 +14,16 @@ class Version701 extends Migration
      */
     public function up()
     {
+        Setting::saveByKey('merchant_StripeV3_enabled','0');
+        Setting::saveByKey('merchant_StripeV3_paymentButtonText','Pay with StripeV3');
+        Setting::saveByKey('merchant_StripeV3_publishableKey','');
+        Setting::saveByKey('merchant_StripeV3_secretKey','');
+        Setting::saveByKey('merchant_Square_enabled','0');
+        Setting::saveByKey('merchant_Square_paymentButtonText','Pay with Square');
+        Setting::saveByKey('merchant_Square_applicationId','');
+        Setting::saveByKey('merchant_Square_accessToken','');
+        Setting::saveByKey('merchant_Square_locationId','');
+
         Setting::saveByKey('version', '7.0.1');
     }