+++ CHANGELOG.md
@@ -7,7 +7,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## 7.0.1
 - add merchant StripeV3
 - fix missing attribute in email password settings
-- set stripe (V2) apiversion to match stripe-php v10.21.0
+- set stripe (V2 deprecated) apiversion to match stripe-php v10.21.0
+- added Squareup hosted quickpay merchant
+- using Square hosted checkout page currently does not support a "cancel" or "back to merchant" redirect.  
+  user option is browser back button or close page , both of which are no help in returning to BT customer page..
 
 ## 7.0.0
 - Requires PHP "^8.2"