+++ app/Modules/Documentation/Views/documentation/Release-Notes.md
@@ -5,7 +5,8 @@ Release Notes
 ### BillingTrack 7.0.1
 - add merchant StripeV3
 - fix missing attribute in email password settings
-- set stripe (V2) apiversion to match stripe-php v10.21.0
+- set stripe (V2 deprecated) apiversion to match stripe-php v10.21.0
+- added Squareup hosted quickpay merchant
 
 ### BillingTrack 7.0.0
 **Upgrade Existing v6.x.x installation to v7.0.0**